server/requirements.txt
# General Dependencies
requests
python-dotenv

# Logging and Utilities
loguru

# Task Queue and Message Broker
redis

# RAG
llama-hub==0.0.45
llama-index==0.8.66
qdrant_client
nltk
sentence-transformers

# Storage
minio==7.1.17

# Ray
ray
ray[data]

# Reader Handling
pypdf
pyarrow
Jinja2
Pillow
wikipedia
nbconvert

# Data Validation
pydantic

# Http Server
fastapi
uvicorn